minute bug fix
glorious eggroll had a bug where it would always use the tag value
instead of only when -t was passed in. This mistake has now been
fixed.

This was from adding in the tag pass in, then I added in the datafile
to be passed as first argument, bumping tflag to second argument. The
check statement was not updated to look for this